refactor(base-gen): extract ExtractionRunLifecycleMixin (Tier-1b Task 5)
Move the four concrete extraction-run lifecycle methods
(_execute_extractor, _fail_active_extraction_runs,
_finalize_extraction_runs, _mark_extraction_runs_finalization_failed)
verbatim from BaseGenerationService into a new
base_generation/_extraction_lifecycle.py ExtractionRunLifecycleMixin,
composed ahead of the existing mixins and ABC.

Behavior-preserving: bodies moved byte-for-byte, including the
isinstance(result, ExtractionOutcome) write-guard that gates the
_last_token_totals / _last_extraction_run_ids billing accumulators
(the double-bill idempotency backstop) and the finalize-vs-fail
atomicity path. ExtractorExecutionError / EXTRACTOR_TIMEOUT_SECONDS are
imported function-locally to avoid a circular import; base-owned
attributes and abstract methods called via self get annotation-only
stubs. __abstractmethods__ frozenset unchanged (8).
