+++ reflexio/server/services/base_generation/__init__.py
@@ -10,6 +10,9 @@ FIFO drain, ``_run_generation`` billing emit-ordering, and the lock/queue helper
 from reflexio.server.services.base_generation._config_filter import (
     ConfigFilterMixin,
 )
+from reflexio.server.services.base_generation._extraction_lifecycle import (
+    ExtractionRunLifecycleMixin,
+)
 from reflexio.server.services.base_generation._should_run import (
     ShouldRunPrecheckMixin,
 )
@@ -20,6 +23,7 @@ from reflexio.server.services.base_generation._status_change import (
 
 __all__ = [
     "ConfigFilterMixin",
+    "ExtractionRunLifecycleMixin",
     "ShouldRunPrecheckMixin",
     "StatusChangeMixin",
     "StatusChangeOperation",
